pyproject.toml
[project]
name = "recursivegarden"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-vite>=3.0.6",
    "djangorestframework>=3.15.2",
    "djangorestframework-api-key>=3.0.0",
    "djlint>=1.36.4",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "ruff>=0.9.6",
    "slippers>=0.6.2",
]
